make the max cache size consistent to long

+++ xmppserver/src/main/java/org/jivesoftware/util/cache/Cache.java
@@ -83,7 +83,7 @@ public interface Cache<K extends Serializable, V extends Serializable> extends j
      *
      * @param maxSize the maximum size of the cache in bytes.
      */
-    void setMaxCacheSize(int maxSize);
+    void setMaxCacheSize(long maxSize);
 
     /**
      * Returns the maximum number of milliseconds that any object can live
@@ -116,7 +116,7 @@ public interface Cache<K extends Serializable, V extends Serializable> extends j
      *
      * @return the size of the cache contents in bytes.
      */
-    int getCacheSize();
+    long getCacheSize();
 
     /**
      * Returns the number of cache hits. A cache hit occurs every
